+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * https support for server-side reports
  * Fixed chart HTML4/5 rendering
  * Fixed invalid PDF missing image placeholder format
+ * Fixed InvalidImage rendering in non-WinForms applications
 
 # 15.1.13
  * Fixed BinaryFormatter/ResX issues with PDF export on .NET 5